sigpack: Add signed package scripts
These package/unpackage scripts will be used for End-to-End
security. Host machine can sign given package(container image)
and encrypt it for a specific client (if needed).
Client can verify the signed package (and decrypt it if needed).

This ensures that the container image is sane image from the
authorized host machine, and also host machine ensures that
the image is only for a specific client.

Maybe, it is better to integrate it to crypto on squashfs.