ermine-breeder: Fix cleanup path not to be error
In shellscript, [ <cond> ] && <do> is usually used as
an shorter alternative of if ... ;then ... fi. But
the former state can return an error if it is used
in the end of function. In that case, since the function
returns an error, the shellscript with set -e, will
unexpectedly exit.

Anyway, rm -rf DIR forcibly remove DIR and return success.
we don't need to check the DIR exist or not.