ermine: minc-moult: um: Hide console messages in minc mode
Hide kernel console messages in minc mode with user-mode linux.
For qemu mode, host console is connected to its serial, so all
messages on ttyS0 will be output, but tty0 is hidden.
In user-mode linux, host console is connected to tty0. So to
hide kernel messages from user, we have to output it into
ttyS0.