minc-dens: Suppress error message of grep
Suppress errors if there is no lease file at searching
unused ip, because after clean install, it must not
be there.

Note that we ignores errors when releasing ip address,
but do not suppress the error message because it is
not normal situation. Some problem should be there
(e.g. unexpectedly removed the lease file)