minc: Redirect /dev/console to tty in container
Redirect container stdout and stderr to /dev/console
and redirect it again to current tty.
In the container tty command will always show /dev/console.

This also unifies all output to stderr and stdout of
container command to $TEMPDIR/log.
Note that --debug messages are written in $TEMPDIR/debug.

And also, a small fix to remove echoing Exit report
when background mode enables, because it always
breaks output of container commands.