mount overlayfs correctly on ubuntu 16.04
This part is workaround for ubuntu 14.10 or earlier when the module
name of overlayfs was "overlayfs". However, ubuntu 16.04 has the name
"overlayfs" as an alias for the module name. So, /proc/filesystems has
both "overlay" and "overlayfs", and the mount without "workdir" option
is executed and the mount fails.

This patch prevents the mount without workdir on Ubuntu 16.04, and
mounts it correctly.